requirements.txt
anthropic>=0.18.0
streamlit>=1.32.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
nba-api>=1.4.1
requests>=2.31.0
beautifulsoup4>=4.12.0
scikit-learn>=1.4.0
xgboost>=2.0.0
pandas>=2.2.0
numpy>=1.26.0
pytest>=8.0.0
mcp>=1.0.0
